requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
python-dotenv
cloudscraper
websockets
PyJWT
